fahrenheit de vergulden sonra bir 0, selside ise iki sifir qalir?

+++ src/main/java/Lesson8/MinutToYear.java
@@ -7,9 +7,9 @@ public class MinutToYear {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter the number of minutes");
         int minut = scanner.nextInt();
-        double year = minut/525600;
-        int days = (int) (minut/60/24)%365;
-        System.out.printf("%d is %f years and %d days", minut, year, days);
+        int year = (int) (minut / 525600);
+        int days = (int) (minut / 60 / 24) % 365;
+        System.out.printf("%f is %f years and %d days", minut, year, days);
 
 
     }
